Authentication added

+++ CriticAPI/game.py
@@ -33,22 +33,52 @@ def API3(request):
 
 def getList(request):
     statusCode = 200
+    result = "bad"
+    content_type = None
+
+    if "Authorization" in request.headers:
+        auth = universal.decode_token(request.headers["Authorization"])
+        if auth[0] == False:
+            return HttpResponse (result, content_type, 401)
+    else:
+        return HttpResponse (result, content_type, 401)
+
     with connection.cursor() as cursor:
         cursor.execute("SELECT * FROM public.game")
         row = universal.dictfetchall(cursor)
-    result = universal.dumpJson(row)
+    result = universal.dumpJson(row)  
     return HttpResponse (result, status = statusCode, content_type = "application/json")
 
 def getGame(request, index1):
     statusCode = 200
+    result = "bad"
+    content_type = None
+
+    if "Authorization" in request.headers:
+        auth = universal.decode_token(request.headers["Authorization"])
+        if auth[0] == False:
+            return HttpResponse (result, content_type, 401)
+    else:
+        return HttpResponse (result, content_type, 401)
+
     with connection.cursor() as cursor:
         cursor.execute("SELECT * FROM public.game WHERE id = %s", [index1])
         row = universal.dictfetchall(cursor)
-    result = universal.dumpJson(row)   
+    result = universal.dumpJson(row)      
     return HttpResponse (result, status = statusCode, content_type = "application/json")
 
 def getCategoryList(request, index1):
     statusCode = 200
+    result = "bad"
+    content_type = None
+
+    if "Authorization" in request.headers:
+        auth = universal.decode_token(request.headers["Authorization"])
+        if auth[0] == False:
+            return HttpResponse (result, content_type, 401)
+    else:
+        return HttpResponse (result, content_type, 401)
+
     with connection.cursor() as cursor:
         cursor.execute("SELECT * FROM public.game WHERE category_id = %s", [index1])
         row = universal.dictfetchall(cursor)
@@ -56,36 +86,92 @@ def getCategoryList(request, index1):
     return HttpResponse (result, status = statusCode, content_type = "application/json")
 
 def create(request):
+    statusCode = 200
+    result = "bad"
+    content_type = None
+
+    if "Authorization" in request.headers:
+        auth = universal.decode_token(request.headers["Authorization"])
+        if auth[0] == False:
+            return HttpResponse (result, content_type, 401)
+    else:
+        return HttpResponse (result, content_type, 401)  
+
     input = universal.getText(request.body)
-    if input == False:
-        return HttpResponse ("Error", status = 400)
-    if "name" not in input:
-        return HttpResponse ("Error", status = 400)  
-    statusCode = 201
+    body = input[1]
+    if input[0] == False:
+        return HttpResponse ("ErrorA", status = 400)
+    if "name" not in body:
+        return HttpResponse ("ErrorB", status = 400)
+        
+    scope = request.headers["scope"]
+    scope = int(scope)
+    if scope != 2:
+        return HttpResponse("Not admin", content_type, 403)         
+    statusCode = 201  
     with connection.cursor() as cursor:
-        cursor.execute("INSERT INTO public.game(category_id, name, price, description) VALUES (%s, %s, %s, %s) RETURNING id, name", [input["category_id"], input["name"], input["price"], input["description"]])
+        cursor.execute("INSERT INTO public.game(category_id, name, price, description) VALUES (%s,%s,%s,%s) RETURNING id, name", [body["category_id"], body["name"], body["price"], body["description"]])
         returnedId = universal.dictfetchall(cursor)
     result = universal.dumpJson(returnedId)
     return HttpResponse (result, status = statusCode, content_type = "application/json")
 
 def update(request, index1):
-    input = universal.getText(request.body)
-    if input == False:
-        return HttpResponse ("Error", status = 400)
-    if "name" not in input:
-        return HttpResponse ("Error", status = 400)      
     statusCode = 200
+    result = "bad"
+    content_type = None
+
+    if "Authorization" in request.headers:
+        auth = universal.decode_token(request.headers["Authorization"])
+        if auth[0] == False:
+            return HttpResponse (result, content_type, 401)
+    else:
+        return HttpResponse (result, content_type, 401)   
+
+    input = universal.getText(request.body)
+    body = input[1]
+    if input == False:
+        return HttpResponse ("ErrorA", status = 400)
+    if "name" not in body:
+        return HttpResponse ("ErrorB", status = 400)      
+    statusCode = 200
+    result = "bad"
+    content_type = None
+
+    if "Authorization" in request.headers:
+        auth = universal.decode_token(request.headers["Authorization"])
+        if auth[0] == False:
+            return HttpResponse (result, content_type, 401)
+    else:
+        return HttpResponse (result, content_type, 401)
+
     with connection.cursor() as cursor:
-        cursor.execute("UPDATE public.game SET name = %s, price = %s, description = %s WHERE id = %s", [input["name"], input["price"], input["description"], index1])
+        cursor.execute("UPDATE public.game SET name = %s, price = %s, description = %s WHERE id = %s", [body["name"], body["price"], body["description"], index1])
     result = "Zaidimas atnaujintas"
     return HttpResponse(result, status = statusCode)
 
 def delete(request, index1):
     statusCode = 200
+    result = "bad"
+    content_type = None
+
+    if "Authorization" in request.headers:
+        auth = universal.decode_token(request.headers["Authorization"])
+        if auth[0] == False:
+            result = "Unauthorized"
+            return HttpResponse (result, content_type, 401)
+    else:
+        result = "No token"
+        return HttpResponse (result, content_type, 401)   
+
+    scope = request.headers["scope"]
+    scope = int(scope)
+    if scope != 2:
+        return HttpResponse("Not admin" + scope, content_type, 403)        
+
     with connection.cursor() as cursor:
         cursor.execute("SELECT * FROM public.review WHERE game_id = %s", [index1])
         row = universal.dictfetchall(cursor)
         if len(row) >= 1:
             return HttpResponse("Sis zaidimas turi apzvalgu.", status = 409)
         cursor.execute("DELETE FROM public.game WHERE id = %s", [index1])        
-    return HttpResponse(status = statusCode)
+    return HttpResponse(status = statusCode)
